Examen recuperació

+++ composeApp/src/androidMain/kotlin/cat/itb/m78/exercices/mapScreen/MapScreenViewModel.kt
@@ -16,7 +16,7 @@ class MapScreenViewModel : ViewModel() {
 
     fun addMarker(lati: String, long: String, navigateToEditMarker : (Long) -> Unit) {
         // Examen recuperació
-        markersTable.create({lati.toLong()-0.01}.toString(), long)
+        markersTable.create({lati.toDouble()-0.01}.toString(), long)
         //
         val id = markersTable.readIdByLatLng(lati, long).executeAsOne()
         navigateToEditMarker(id)
